# ROX-28900: Remove monaco and swap to plain codeblock for code viewer