--- a/ui/apps/platform/package.json
+++ b/ui/apps/platform/package.json
@@ -72,6 +72,7 @@
         "react-router-hash-link": "^2.4.3",
         "react-select": "^2.0.0",
         "react-spinners": "^0.10.4",
+        "react-syntax-highlighter": "^15.6.1",
         "react-table-6": "^6.11.0",
         "react-toastify": "^7.0.3",
         "react-transition-group": "^4.4.1",