--- a/ui/apps/platform/src/Components/CodeViewer.tsx
+++ b/ui/apps/platform/src/Components/CodeViewer.tsx
@@ -1,39 +1,77 @@
-import React, { CSSProperties, ReactNode, useState } from 'react';
-import {
-    CodeBlockAction,
-    ClipboardCopyButton,
-    Button,
-    CodeBlock,
-    CodeBlockCode,
-} from '@patternfly/react-core';
+import React, {
+    createContext,
+    CSSProperties,
+    Dispatch,
+    ReactNode,
+    SetStateAction,
+    useContext,
+    useState,
+} from 'react';
+import { CodeBlockAction, ClipboardCopyButton, Button, CodeBlock } from '@patternfly/react-core';
 import { MoonIcon, SunIcon } from '@patternfly/react-icons';
+
+import { PrismLight as SyntaxHighlighter } from 'react-syntax-highlighter';
+import lightTheme from 'react-syntax-highlighter/dist/esm/styles/prism/one-light';
+import darkTheme from 'react-syntax-highlighter/dist/esm/styles/prism/one-dark';
+import yaml from 'react-syntax-highlighter/dist/esm/languages/prism/yaml';
 
 import useClipboardCopy from 'hooks/useClipboardCopy';
 
+SyntaxHighlighter.registerLanguage('yaml', yaml);
+
+const CodeViewerThemeContext = createContext<
+    ['light' | 'dark', Dispatch<SetStateAction<'light' | 'dark'>>] | undefined
+>(undefined);
+
+export const CodeViewerThemeProvider = ({ children }) => {
+    const [state, setState] = useState<'light' | 'dark'>('light');
+
+    return (
+        <CodeViewerThemeContext.Provider value={[state, setState]}>
+            {children}
+        </CodeViewerThemeContext.Provider>
+    );
+};
+
+export const useCodeViewerThemeContext = () => {
+    const context = useContext(CodeViewerThemeContext);
+    // Fallback state provides the ability to toggle theme for a single instance if no provider is detected uptree
+    const fallbackState = useState<'light' | 'dark'>('light');
+    return context ?? fallbackState;
+};
+
+// When adding to the supported languages, the correct language definition must be imported and registered as well
+type SupportedLanguages = 'yaml';
+
+const defaultStyle = {
+    '--pf-v5-u-max-height--MaxHeight': '300px',
+    '--pf-v5-c-code-block__content--PaddingTop': '0',
+    '--pf-v5-c-code-block__content--PaddingBottom': '0',
+    '--pf-v5-c-code-block__content--PaddingLeft': '0',
+    '--pf-v5-c-code-block__content--PaddingRight': '0',
+    overflowY: 'auto',
+} as const;
+
 export type CodeViewerProps = {
     code: string;
+    language?: SupportedLanguages;
     className?: string;
     style?: CSSProperties;
     additionalControls?: ReactNode;
 };
 
-const defaultStyle = {
-    '--pf-v5-u-max-height--MaxHeight': '300px',
-    overflowY: 'scroll',
-} as const;
-
 export default function CodeViewer({
     code,
-    className = 'pf-v5-u-max-height',
+    language = 'yaml',
+    className = '',
     style,
     additionalControls,
 }: CodeViewerProps) {
     const { wasCopied, setWasCopied, copyToClipboard } = useClipboardCopy();
-
-    const [isDarkMode, setIsDarkMode] = useState(false);
+    const [theme, setTheme] = useCodeViewerThemeContext();
 
     function toggleDarkMode() {
-        setIsDarkMode((prevValue) => !prevValue);
+        setTheme((prevValue) => (prevValue === 'light' ? 'dark' : 'light'));
     }
 
     const actions = (
@@ -54,8 +92,8 @@ export default function CodeViewer({
             <CodeBlockAction>
                 <Button
                     variant="plain"
-                    aria-label={isDarkMode ? 'Set light mode' : 'Set dark mode'}
-                    icon={isDarkMode ? <SunIcon /> : <MoonIcon />}
+                    aria-label={theme === 'light' ? 'Set dark mode' : 'Set light mode'}
+                    icon={theme === 'light' ? <MoonIcon /> : <SunIcon />}
                     onClick={() => toggleDarkMode()}
                 />
             </CodeBlockAction>
@@ -63,14 +101,25 @@ export default function CodeViewer({
         </>
     );
 
+    // TODO - When Tailwind is removed, we likely need to get rid of this font size override
     return (
         <CodeBlock
-            className={`${isDarkMode ? 'pf-v5-theme-dark' : ''} ${className}`}
+            className={`${theme === 'light' ? '' : 'pf-v5-theme-dark'} pf-v5-u-p-0 pf-v5-u-font-size-xs pf-v5-u-max-height ${className}`}
             style={{ ...defaultStyle, ...style }}
             actions={actions}
         >
-            {/* TODO - Once Tailwind is gone we will need to remove this font size override */}
-            <CodeBlockCode className="pf-v5-u-font-size-xs">{code}</CodeBlockCode>
+            <SyntaxHighlighter
+                language={language}
+                showLineNumbers
+                wrapLongLines
+                style={theme === 'light' ? lightTheme : darkTheme}
+                customStyle={{
+                    margin: 0,
+                    background: 'var(--pf-v6-c-code-block--BackgroundColor)',
+                }}
+            >
+                {code}
+            </SyntaxHighlighter>
         </CodeBlock>
     );
 }