--- a/ui/apps/platform/src/Containers/NetworkGraph/common/NetworkPolicies.tsx
+++ b/ui/apps/platform/src/Containers/NetworkGraph/common/NetworkPolicies.tsx
@@ -12,13 +12,12 @@ import {
     EmptyStateHeader,
 } from '@patternfly/react-core';
 import { SelectOption } from '@patternfly/react-core/deprecated';
-import { CodeEditor, Language } from '@patternfly/react-code-editor';
 
 import download from 'utils/download';
 import SelectSingle from 'Components/SelectSingle';
 import useFetchNetworkPolicies from 'hooks/useFetchNetworkPolicies';
 import { getAxiosErrorMessage } from 'utils/responseErrorUtils';
-import CodeEditorDarkModeControl from 'Components/PatternFly/CodeEditorDarkModeControl';
+import CodeViewer from '../../../Components/CodeViewer';
 
 type NetworkPoliciesProps = {
     entityName: string;
@@ -35,7 +34,6 @@ const allNetworkPoliciesId = 'All network policies';
 function NetworkPolicies({ entityName, policyIds }: NetworkPoliciesProps): React.ReactElement {
     const { networkPolicies, networkPolicyErrors, isLoading, error } =
         useFetchNetworkPolicies(policyIds);
-    const [customDarkMode, setCustomDarkMode] = React.useState(false);
 
     const allNetworkPoliciesYAML = useMemo(
         () => ({
@@ -52,10 +50,6 @@ function NetworkPolicies({ entityName, policyIds }: NetworkPoliciesProps): React
     useEffect(() => {
         setSelectedNetworkPolicy(allNetworkPoliciesYAML);
     }, [allNetworkPoliciesYAML]);
-
-    function onToggleDarkMode() {
-        setCustomDarkMode((prevValue) => !prevValue);
-    }
 
     function handleSelectedNetworkPolicy(_, value: string) {
         if (value !== allNetworkPoliciesId) {
@@ -164,21 +158,7 @@ function NetworkPolicies({ entityName, policyIds }: NetworkPoliciesProps): React
                     <>
                         <StackItem>
                             <div className="pf-v5-u-h-100">
-                                <CodeEditor
-                                    isDarkTheme={customDarkMode}
-                                    customControls={
-                                        <CodeEditorDarkModeControl
-                                            isDarkMode={customDarkMode}
-                                            onToggleDarkMode={onToggleDarkMode}
-                                        />
-                                    }
-                                    isCopyEnabled
-                                    isLineNumbersVisible
-                                    isReadOnly
-                                    code={selectedNetworkPolicy.yaml}
-                                    language={Language.yaml}
-                                    height="300px"
-                                />
+                                <CodeViewer code={selectedNetworkPolicy.yaml} />
                             </div>
                         </StackItem>
                         <StackItem>