--- a/ui/apps/platform/src/Containers/NetworkGraph/simulation/NetworkPoliciesYAML.css
+++ b/ui/apps/platform/src/Containers/NetworkGraph/simulation/NetworkPoliciesYAML.css
@@ -1,8 +1,0 @@
-.network-policies-yaml .pf-v5-c-code-editor__controls {
-  width: 100%;
-}
-
-
-.compare-yaml-modal {
-  --pf-v5-c-modal-box--MaxWidth: 1200px;
-}