… ui/apps/platform/src/Components/PatternFly/CodeEditorDarkModeControl.tsx
… ui/apps/platform/src/index.tsx