### Description

Somewhat of a prerequisite for moving from CRA -> Vite, but a beneficial change regardless.

This removes all `monaco` related dependencies and `@patternfly/react-code-editor` in favor of the simple `CodeBlock` that ships will PF core and the lightweight `react-syntax-highlighter` (small, MIT licensed). 

Why? Monaco is L A R G E. So large that it causes OOM when attempting to build with Vite. Monaco is the open source component that drives the full code editor portion of VSCode, and our use in ACS is limited to displaying a read-only view of YAML with syntax highlighting.

Here is a comparison of the webpack build output folder.

Before (total 2.27 MB):

```
  1.24 MB    build/static/js/main.18663eae.js
  312.43 kB  build/static/js/8436.0de22991.chunk.js
  160.14 kB  build/static/js/601.c54973f7.chunk.js
  152.42 kB  build/static/js/3460.310a9f1b.chunk.js
  106.55 kB  build/static/css/main.31d72a89.css
  100.18 kB  build/static/js/6381.e8919607.chunk.js
  98.29 kB   build/json.worker.js
  64.91 kB   build/static/js/5970.d3c4acf0.chunk.js
  63.46 kB   build/editor.worker.js
  44.24 kB   build/static/js/2170.eed13c74.chunk.js
  ...
```

After (total 1.40 MB):

```
  507.18 kB  build/static/js/main.ca46b6c0.js
  312.43 kB  build/static/js/8436.0de22991.chunk.js
  160.77 kB  build/static/js/3524.d3d32866.chunk.js
  152.42 kB  build/static/js/3460.310a9f1b.chunk.js
  100.18 kB  build/static/js/6381.e8919607.chunk.js
  93.09 kB   build/static/css/main.08595755.css
  63.93 kB   build/static/js/3443.0e474573.chunk.js
  44.63 kB   build/static/js/3109.20b0ac34.chunk.js
  40.88 kB   build/static/js/5579.359da093.chunk.js
  39.65 kB   build/static/js/47.5f666508.chunk.js
  ...
```

This is an overall reduction of about 37% in the total disk size of assets the ACS UI ships. Of note, quite a bit of this was reduction is in our **main.js** file, which needs to be downloaded and blocks the page upon every first visit to the app. The reduction in this file specifically was 1,240 kB -> 507 kB (almost 60% smaller).

As a bonus, our production build time has also decreased:
```
before: npm run build  528.05s user 75.57s system 168% cpu 5:58.35 total
after:  npm run build  326.91s user 48.64s system 168% cpu 3:42.62 total
```

I suspect that dev build times are also quicker, but did not benchmark them.

Visual comparison of old and new light theme:
![image](https://github.com/user-attachments/assets/d8b10f0c-400e-4523-8b2c-7c90d2edc0f7)
![image](https://github.com/user-attachments/assets/1f754418-d5aa-4e73-a2a7-15faaa70646a)


Visual comparison of old and new dark theme:
![image](https://github.com/user-attachments/assets/72d8acd2-c1af-4866-9cf1-b3abf17089fa)
![image](https://github.com/user-attachments/assets/e49bfca9-4c04-4bde-904c-a8d325ba5695)

The most obvious visual change is the right-alignment of code block icons. I did not left align these as there was no out-of-the box way to do so with PatternFly's CodeBlock component and doing so would be arbitrarily going against the PF design.


### User-facing documentation

- [ ] CHANGELOG is updated **OR** update is not needed
- [ ] [documentation PR](https://spaces.redhat.com/display/StackRox/Submitting+a+User+Documentation+Pull+Request) is created and is linked above **OR** is not needed

### Testing and quality

<!--
General Availability requirements: https://github.com/stackrox/stackrox/blob/master/PR_GA.md
Feature Flags usage: https://github.com/stackrox/stackrox/blob/master/pkg/features/README.md
-->

- [ ] the change is production ready: the change is GA or otherwise the functionality is gated by a feature flag
- [ ] CI results are inspected

#### Automated testing

<!--
If no tests have been contributed, please explain why unless it's obvious,
e.g., the PR is a one-line comment change.
-->

- [ ] added unit tests
- [ ] added e2e tests
- [ ] added regression tests
- [ ] added compatibility tests
- [ ] modified existing tests

#### How I validated my change

Added component tests to cover basic functionality, copy to clipboard, and theme changing.

Manual test of Network Graph policy gen:
![image](https://github.com/user-attachments/assets/27b9f657-a62a-4bbe-9f2d-72348a235c8f)
![image](https://github.com/user-attachments/assets/e20eae37-0cbf-4823-ae59-19e86fb82d41)

Manual test of Violation net policies:
![image](https://github.com/user-attachments/assets/816b46ba-ef7b-46cd-8452-a9658e56ecf9)